+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -1,0 +1,8 @@
+package wooteco.subway.dao;
+
+import wooteco.subway.domain.Section;
+
+public interface SectionDao {
+
+    Long save(Section section);
+}
