+++ src/main/java/wooteco/subway/dao/SectionDaoImpl.java
@@ -1,0 +1,36 @@
+package wooteco.subway.dao;
+
+import java.sql.PreparedStatement;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import wooteco.subway.domain.Section;
+
+@Repository
+public class SectionDaoImpl implements SectionDao{
+
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public SectionDaoImpl(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public Long save(Section section) {
+        final String sql = "INSERT INTO section (line_id, up_station_id, down_station_id, distance) VALUES (?, ?, ?, ?)";
+
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setLong(1, section.getLineId());
+            ps.setLong(2, section.getUpStationId());
+            ps.setLong(3, section.getDownStationId());
+            ps.setInt(4, section.getDistance());
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
+    }
+}
