+++ src/main/java/wooteco/subway/dao/entity/StationEntity.java
@@ -1,4 +1,4 @@
-package wooteco.subway.entity;
+package wooteco.subway.dao.entity;
 
 import java.util.Objects;
