+++ src/main/java/wooteco/subway/domain/Section.java
@@ -1,0 +1,31 @@
+package wooteco.subway.domain;
+
+public class Section {
+    private final Long lineId;
+    private final Long upStationId;
+    private final Long downStationId;
+    private final int distance;
+
+    public Section(Long lineId, Long upStationId, Long downStationId, int distance) {
+        this.lineId = lineId;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+
+    public Long getLineId() {
+        return lineId;
+    }
+}
