+++ src/main/java/wooteco/subway/service/StationService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import wooteco.subway.dao.StationDao;
+import wooteco.subway.dao.entity.StationEntity;
 import wooteco.subway.domain.Station;
-import wooteco.subway.entity.StationEntity;
 import wooteco.subway.service.dto.station.StationFindResponse;
 import wooteco.subway.service.dto.station.StationSaveRequest;
 import wooteco.subway.service.dto.station.StationSaveResponse;
