+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -14,25 +14,24 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.ui.dto.LineResponse;
+import wooteco.subway.ui.dto.StationResponse;
 
-@DisplayName("지하철 노선 관련 기능")
-class LineEntityAcceptanceTest extends AcceptanceTest {
+@DisplayName("지하철역 관련 기능")
+public class StationAcceptanceTest extends AcceptanceTest {
 
     @Test
-    @DisplayName("존재하지 않는 노선을 생성한다.")
-    void createLine() {
+    @DisplayName("지하철역을 생성한다.")
+    void createStation() {
         // given
         Map<String, String> params = new HashMap<>();
-        params.put("name", "3호선");
-        params.put("color", "bg-orange-600");
+        params.put("name", "선릉역");
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
             .body(params)
             .contentType(MediaType.APPLICATION_JSON_VALUE)
             .when()
-            .post("/lines")
+            .post("/stations")
             .then().log().all()
             .extract();
 
@@ -42,17 +41,16 @@ class LineEntityAcceptanceTest extends AcceptanceTest {
     }
 
     @Test
-    @DisplayName("이미 존재하는 노선을 생성할 수 없다.")
-    void createLineWithDuplicateName() {
+    @DisplayName("기존에 존재하는 지하철역 이름으로 지하철역을 생성한다.")
+    void createStationWithDuplicateName() {
         // given
         Map<String, String> params = new HashMap<>();
-        params.put("name", "4호선");
-        params.put("color", "bg-blue-600");
+        params.put("name", "강남역");
         RestAssured.given().log().all()
             .body(params)
             .contentType(MediaType.APPLICATION_JSON_VALUE)
             .when()
-            .post("/lines")
+            .post("/stations")
             .then().log().all()
             .extract();
 
@@ -61,7 +59,7 @@ class LineEntityAcceptanceTest extends AcceptanceTest {
             .body(params)
             .contentType(MediaType.APPLICATION_JSON_VALUE)
             .when()
-            .post("/lines")
+            .post("/stations")
             .then()
             .log().all()
             .extract();
@@ -71,35 +69,33 @@ class LineEntityAcceptanceTest extends AcceptanceTest {
     }
 
     @Test
-    @DisplayName("노선들을 조회한다.")
-    void getLines() {
-        // given
+    @DisplayName("지하철역을 조회한다.")
+    void getStations() {
+        /// given
         Map<String, String> params1 = new HashMap<>();
-        params1.put("name", "1호선");
-        params1.put("color", "bg-blue-600");
+        params1.put("name", "강남역");
         ExtractableResponse<Response> createResponse1 = RestAssured.given().log().all()
             .body(params1)
             .contentType(MediaType.APPLICATION_JSON_VALUE)
             .when()
-            .post("/lines")
+            .post("/stations")
             .then().log().all()
             .extract();
 
         Map<String, String> params2 = new HashMap<>();
-        params2.put("name", "2호선");
-        params2.put("color", "bg-green-600");
+        params2.put("name", "역삼역");
         ExtractableResponse<Response> createResponse2 = RestAssured.given().log().all()
             .body(params2)
             .contentType(MediaType.APPLICATION_JSON_VALUE)
             .when()
-            .post("/lines")
+            .post("/stations")
             .then().log().all()
             .extract();
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
             .when()
-            .get("/lines")
+            .get("/stations")
             .then().log().all()
             .extract();
 
@@ -108,24 +104,23 @@ class LineEntityAcceptanceTest extends AcceptanceTest {
         List<Long> expectedLineIds = Arrays.asList(createResponse1, createResponse2).stream()
             .map(it -> Long.parseLong(it.header("Location").split("/")[2]))
             .collect(Collectors.toList());
-        List<Long> resultLineIds = response.jsonPath().getList(".", LineResponse.class).stream()
-            .map(LineResponse::getId)
+        List<Long> resultLineIds = response.jsonPath().getList(".", StationResponse.class).stream()
+            .map(it -> it.getId())
             .collect(Collectors.toList());
         assertThat(resultLineIds).containsAll(expectedLineIds);
     }
 
     @Test
-    @DisplayName("존재하는 노선을 제거한다. 상태코드는 200 이어야 한다.")
+    @DisplayName("존재하는 지하철역을 제거한다. 상태코드는 200 이어야 한다.")
     void deleteStation() {
         // given
         Map<String, String> params = new HashMap<>();
-        params.put("name", "1호선");
-        params.put("color", "bg-blue-600");
+        params.put("name", "강남역");
         ExtractableResponse<Response> createResponse = RestAssured.given().log().all()
             .body(params)
             .contentType(MediaType.APPLICATION_JSON_VALUE)
             .when()
-            .post("/lines")
+            .post("/stations")
             .then().log().all()
             .extract();
 
@@ -142,60 +137,11 @@ class LineEntityAcceptanceTest extends AcceptanceTest {
     }
 
     @Test
-    @DisplayName("존재하지 않는 노선을 제거한다. 상태코드는 204 이어야 한다.")
+    @DisplayName("존재하지 않는 지하철역을 제거한다. 상태코드는 204 이어야 한다.")
     void deleteNonStation() {
         RestAssured.given().log().all()
             .when()
-            .delete("lines/1")
-            .then().log().all()
-            .statusCode(HttpStatus.NO_CONTENT.value());
-    }
-
-    @Test
-    @DisplayName("존재하는 노선을 수정한다. 상태코드는 200이어야 한다.")
-    void updateLine() {
-        // given
-        Map<String, String> params1 = new HashMap<>();
-        params1.put("name", "1호선");
-        params1.put("color", "bg-red-600");
-        ExtractableResponse<Response> createResponse = RestAssured.given().log().all()
-            .body(params1)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when()
-            .post("/lines")
-            .then().log().all()
-            .extract();
-
-        Map<String, String> params2 = new HashMap<>();
-        params2.put("name", "2호선");
-        params2.put("color", "bg-green-600");
-
-        // when
-        String uri = createResponse.header("Location");
-
-        // then
-        RestAssured.given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(params2)
-            .when()
-            .put(uri)
-            .then().log().all()
-            .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 노선을 수정한다. 상태코드는 204이어야 한다.")
-    void updateNonLine() {
-        // given
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "1호선");
-        params.put("color", "bg-red-600");
-
-        RestAssured.given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(params)
-            .when()
-            .put("/lines/1")
+            .delete("stations/1")
             .then().log().all()
             .statusCode(HttpStatus.NO_CONTENT.value());
     }
