+++ src/test/java/wooteco/subway/dao/SectionDaoImplTest.java
@@ -1,0 +1,29 @@
+package wooteco.subway.dao;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.domain.Section;
+
+@JdbcTest
+public class SectionDaoImplTest {
+
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+    private SectionDao sectionDao;
+
+    @BeforeEach
+    void setUp() {
+         sectionDao = new SectionDaoImpl(jdbcTemplate);
+    }
+
+    @Test
+    void save() {
+        Section section = new Section(1L, 2L, 3L, 10);
+        assertDoesNotThrow(() -> sectionDao.save(section));
+    }
+}
